Fix flaky tests by preventing Safari optimisaiton.
Safari would optimise the `while` loop in the `sleep` function,
such that it didn't hold up the main thread for the given duration.
Changing the duration slightly between calls with `Math.random()`
prevents Safari's optimisation and ensures the `sleep`
method holds up the main thread for the given number of milliseconds.