Bind `this` in methods that use it
Polyfill helpers use `this` to refer to the global scope, which is true at declaration time since the generated polyfill code explicitly scopes `this` to be the global scope.

However, if this code is later executed in a strit-mode context, it would fail, since `this` is whatever the parent's context is.

This commit makes these four methods strict-mode compliant by binding `this` to its value at declaration time, which is explicitly set to the global scope.