prevent focusin detection from scrolling
in order to prevent focusin from accidentally scrolling chrome, setting it a `position:fixed` does the trick. Yet, as this is not widely used and impacts fewer browsers, I would rather like to removing this from default polyfill and give to the developer the liberty of having this detection run. Modifying the dom is always costly!